Implement allocation_candidates for flavor plugin
This adds the flavor plugin, where users request
a reservation similar to:

  openstack reservation lease create test --reservation \
    resource_type=flavor:instance,amount=2,\
    flavor_id='e26a4241-b83d-4516-8e0e-8ce2665d1966'

This uses lots of ideas from the instance reservation
plugin, and where possible leans on some of the existing
logic, currently only filter_hosts_by_reservation.

The plugin is not currently useable.
There are follow on patches that implement
reserve_resource and then finally update_reservation.

Change-Id: I7507e87908e16b313fd624810d6f43cb73dddafe