Make flavor plugin usable
This implementes everything except update_reservation,
so you can create, use and delete flavor based
reservations.

We are able to re-use the on_start and on_end
implementation from the instance plugin,
along with its cleanup_resources implemenation.

However, the flavors created for each reservation
are different, because we copy the details from
the source flavor, then add in the Blazar specific
resource requests on top. This allows for users
to make use of pinned CPUs, hugepages, and simple
PCI device requests and VGPUs. We do not support
all flavor options, and these flavors will be
rejected.

Change-Id: I5b2c9fded67ae7c593ea50ebef0f42b6c5856324